requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
boto==2.37.0
celery==3.1.18
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9
django-admin-honeypot==0.4.0
django-allauth==0.24.1
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-extensions==1.6.1
django-guardian==1.3.2
django-rest-swagger==0.3.4
django-storages-redux==1.2.3
djangorestframework==3.3.2
djangorestframework-gis==0.10.1
ecdsa==0.13
Fabric==1.10.1
factory-boy==2.6.0
fake-factory==0.5.3
gnureadline==6.3.3
gunicorn==19.3.0
invoke==0.10.1
ipython==4.0.0
kombu==3.0.30
mock==1.3.0
model-mommy==1.2.6
oauthlib==1.0.3
paramiko==1.15.2
Pillow==3.0.0
psycopg2==2.6
pycrypto==2.6.1
python-openid==2.2.5
python3-openid==3.0.9
python-social-auth==0.2.14
pytz==2015.2
PyYAML==3.11
redis==2.10.3
requests==2.9.1
requests-oauthlib==0.6.0
sendgrid==1.6.22
six==1.10.0
static3==0.5.1
yolk==0.4.3
